+++ girder/girder_large_image/web_client/views/index.js
@@ -1,9 +1,11 @@
 import ConfigView from './configView';
 import ImageViewerSelectWidget from './imageViewerSelectWidget';
 import * as imageViewerWidget from './imageViewerWidget';
+import ItemViewWidget from './itemViewWidget';
 
 export {
     ConfigView,
     ImageViewerSelectWidget,
-    imageViewerWidget
+    imageViewerWidget,
+    ItemViewWidget
 };