+++ CHANGELOG.md
@@ -4,12 +4,14 @@
 
 ### Features
 - Added bioformats tile source (#463)
+- Handle OME Tiff files with sub-ifd images (#469)
 
 ### Improvements
 - Expose more internal metadata (#479)
 - Improve how Philips XML internal metadata is reported (#475)
 - Show aperio version in internal metadata (#474)
 - Add css classes to metadata on the item page (#472)
+- The Girder web client exports the ItemViewWidget
 
 ### Bug Fixes
 - Add a reference to updated time to avoid overcaching associated images (#477)